renamed styles to styled, removed redundant css
