+++ src/components/IconComponent/styled.js
@@ -2,7 +2,6 @@ import styled from "styled-components";
 
 export const BackgroundIcon = styled.div`
   position: absolute;
-  top: 0px;
   right: 0px;
 `;
